feat: move source display function in common #65
