+++ packages/common/src/soligare/index.ts
@@ -19,6 +19,4 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 // @index('./*', f => `export * from '${f.path}'`)
-export * from "./enums";
 export * from "./interfaces";
-export * from "./constants";
