+++ packages/frontend/src/app/modules/shared/components/display-sources/display-sources.component.html
@@ -38,7 +38,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
             >({{ EXTERNAL_SOURCE_MAPPING[source.name].licenseLabel }})</a
           ></ng-container
         >
-        <span *ngIf="!last">,</span>
+        <span *ngIf="!last">, </span>
       </ng-container>
     </ng-container>
   </div>
